+++ src/main/java/com/drmangotea/tfmg/TFMG.java
@@ -114,6 +114,9 @@ public class TFMG {
 
 
     public static ResourceLocation asResource(String path) {
+        if (path.contains(":")) {
+            return ResourceLocation.tryParse(path);
+        }
         return ResourceLocation.fromNamespaceAndPath(MOD_ID, path);
     }