+++ src/main/java/com/drmangotea/tfmg/registry/TFMGGuiTextures.java
@@ -38,6 +38,7 @@ public enum TFMGGuiTextures implements ScreenElement {
     VAT_SUPERHEATER("chemical_vat", 112, 98, 9, 9),
     VAT_BAROMETER("chemical_vat",211,3,42,42),
     VAT_BAROMETER_NEEDLE("barometer_needle",0,0,96,528),
+    PLATINUM_ELECTRODE("chemical_vat", 202, 0, 8, 29)
     ;